A little modding to get oam to play nice with storagemanager wrt BRM files.

+++ procmon/processmonitor.cpp
@@ -2567,7 +2567,7 @@ pid_t ProcessMonitor::startProcess(string processModuleType, string processName,
             system(cmd.c_str());
 
             // if Non Parent OAM Module, get the dbmr data from Parent OAM Module
-            if ( !gOAMParentModuleFlag && !HDFS )
+            if ( !gOAMParentModuleFlag && !HDFS && DBRootStorageType != "storagemanager")
             {
 
                 //create temp dbrm directory
@@ -2673,7 +2673,7 @@ pid_t ProcessMonitor::startProcess(string processModuleType, string processName,
                 }
 
                 // now delete the dbrm data from local disk
-                if ( !gOAMParentModuleFlag && !HDFS && DataRedundancyConfig == "n")
+                if ( !gOAMParentModuleFlag && !HDFS && DBRootStorageType != "storagemanager" && DataRedundancyConfig == "n")
                 {
                     string cmd = "rm -f " + DBRMDir + "/*";
                     system(cmd.c_str());
